# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ITIL v4 Change Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /changes:
    get:
      operationId: listChanges
      tags: [changes]
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: track, schema: { type: string } }
        - { in: query, name: ownerId, schema: { type: integer } }
        - { in: query, name: search, schema: { type: string } }
      responses:
        "200":
          description: Changes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChangeRequest" }
    post:
      operationId: createChange
      tags: [changes]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateChangeBody" } } }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/ChangeRequest" } } }
  /changes/{id}:
    get:
      operationId: getChange
      tags: [changes]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Change
          content: { application/json: { schema: { $ref: "#/components/schemas/ChangeDetail" } } }
    patch:
      operationId: updateChange
      tags: [changes]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateChangeBody" } } }
      responses:
        "200":
          description: Updated
          content: { application/json: { schema: { $ref: "#/components/schemas/ChangeRequest" } } }
    delete:
      operationId: deleteChange
      tags: [changes]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /changes/{id}/transition:
    post:
      operationId: transitionChange
      tags: [changes]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/TransitionBody" } } }
      responses:
        "200":
          description: Transitioned
          content: { application/json: { schema: { $ref: "#/components/schemas/ChangeRequest" } } }
        "400": { description: Invalid transition, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  # ---------- TEMPLATES ----------
...
  /changes/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Comment" }
    post:
      operationId: createComment
      tags: [comments]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateCommentBody" } } }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/Comment" } } }

  # ---------- DASHBOARD ----------
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    ChangeRequest:
      type: object
      properties:
        id: { type: integer }
        ref: { type: string }
        title: { type: string }
        description: { type: string }
        track: { type: string, enum: [normal, standard, emergency] }
        status: { type: string }
        risk: { type: string, enum: [low, medium, high] }
        impact: { type: string, enum: [low, medium, high] }
        priority: { type: string, enum: [low, medium, high, critical] }
        category: { type: string }
        ownerId: { type: integer }
        ownerName: { type: string }
        assigneeId: { type: ["integer", "null"] }
        assigneeName: { type: ["string", "null"] }
        templateId: { type: ["integer", "null"] }
        templateName: { type: ["string", "null"] }
        plannedStart: { type: ["string", "null"], format: date-time }
        plannedEnd: { type: ["string", "null"], format: date-time }
        actualStart: { type: ["string", "null"], format: date-time }
        actualEnd: { type: ["string", "null"], format: date-time }
        cabMeetingId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, ref, title, description, track, status, risk, impact, priority, category, ownerId, ownerName, createdAt, updatedAt]
    ChangeDetail:
      allOf:
        - $ref: "#/components/schemas/ChangeRequest"
        - type: object
          properties:
            planning: { $ref: "#/components/schemas/PlanningRecord" }
            testing: { $ref: "#/components/schemas/TestRecord" }
            pir: { $ref: "#/components/schemas/PirRecord" }
            approvals:
              type: array
              items: { $ref: "#/components/schemas/Approval" }
            comments:
              type: array
              items: { $ref: "#/components/schemas/Comment" }
    CreateChangeBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        track: { type: string, enum: [normal, standard, emergency] }
        risk: { type: string, enum: [low, medium, high] }
        impact: { type: string, enum: [low, medium, high] }
        priority: { type: string, enum: [low, medium, high, critical] }
        category: { type: string }
        templateId: { type: ["integer", "null"] }
        assigneeId: { type: ["integer", "null"] }
        plannedStart: { type: ["string", "null"], format: date-time }
        plannedEnd: { type: ["string", "null"], format: date-time }
      required: [title, description, track, risk, impact, priority, category]
    UpdateChangeBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        risk: { type: string, enum: [low, medium, high] }
        impact: { type: string, enum: [low, medium, high] }
        priority: { type: string, enum: [low, medium, high, critical] }
        category: { type: string }
        assigneeId: { type: ["integer", "null"] }
        plannedStart: { type: ["string", "null"], format: date-time }
        plannedEnd: { type: ["string", "null"], format: date-time }
        cabMeetingId: { type: ["integer", "null"] }
    TransitionBody:
      type: object
      properties:
        toStatus: { type: string }
        note: { type: ["string", "null"] }
      required: [toStatus]
...
    PlanningRecord:
      type: object
      properties:
        changeId: { type: integer }
        scope: { type: string }
        implementationPlan: { type: string }
        rollbackPlan: { type: string }
        riskAssessment: { type: string }
        impactedServices: { type: string }
        communicationsPlan: { type: string }
        successCriteria: { type: string }
        signedOff: { type: boolean }
        signedOffAt: { type: ["string", "null"], format: date-time }
        signedOffBy: { type: ["string", "null"] }
      required: [changeId, scope, implementationPlan, rollbackPlan, riskAssessment, impactedServices, communicationsPlan, successCriteria, signedOff]
...
    TestCase:
      type: object
      properties:
        name: { type: string }
        steps: { type: string }
        expectedResult: { type: string }
        actualResult: { type: string }
        status: { type: string, enum: [pending, passed, failed, blocked] }
      required: [name, steps, expectedResult, actualResult, status]
    TestRecord:
      type: object
      properties:
        changeId: { type: integer }
        testPlan: { type: string }
        environment: { type: string }
        testedBy: { type: ["string", "null"] }
        testedAt: { type: ["string", "null"], format: date-time }
        overallResult: { type: string, enum: [pending, passed, failed] }
        notes: { type: string }
        cases:
          type: array
          items: { $ref: "#/components/schemas/TestCase" }
      required: [changeId, testPlan, environment, overallResult, notes, cases]
...
    PirRecord:
      type: object
      properties:
        changeId: { type: integer }
        outcome: { type: string, enum: [successful, partial, failed, rolledback] }
        objectivesMet: { type: string }
        issuesEncountered: { type: string }
        lessonsLearned: { type: string }
        followupActions: { type: string }
        completedBy: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"], format: date-time }
      required: [changeId, outcome, objectivesMet, issuesEncountered, lessonsLearned, followupActions]
...
    Approval:
      type: object
      properties:
        id: { type: integer }
        changeId: { type: integer }
        roleKey: { type: string }
        roleName: { type: string }
        approverId: { type: ["integer", "null"] }
        approverName: { type: ["string", "null"] }
        decision: { type: string, enum: [pending, approved, rejected, abstain] }
        comment: { type: ["string", "null"] }
        decidedAt: { type: ["string", "null"], format: date-time }
        viaDeputy: { type: boolean }
      required: [id, changeId, roleKey, roleName, decision, viaDeputy]
...
    Comment:
      type: object
      properties:
        id: { type: integer }
        changeId: { type: integer }
        authorId: { type: integer }
        authorName: { type: string }
        body: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, changeId, authorId, authorName, body, createdAt]
    CreateCommentBody:
      type: object
      properties:
        body: { type: string }
      required: [body]
...